+++ make.sh
@@ -11,6 +11,7 @@ npm install @vscode/test-cli
 npm install --save-dev sinon @types/sinon
 
 pip install psutil
+# build for compatibility with older Xcode versions
 pyinstaller --onefile src/XCBBuildServiceProxy/XCBBuildService.py
 pyinstaller --onefile src/XCBBuildServiceProxy/SWBBuildService.py